Pendulum1
Consider the simple pendulm.
We’ve just got a mass of m hanging from a string with length L that is swinging back and forth.

It’s basically as simple of a system as we can work with. Don’t let this simplicity fool you though, it can create some interesting dynamics. We’ll use this as a starting point to introduce some control theory and compare that to continuous control reinforcement learning. Before we get to that, we need to spend some time understanding the dynamics of the system and how to simulate it!